feat: add board config and staff modules
